+++ src/app/Scenes/Home/Components/NewWorksForYouRail.tsx
@@ -70,7 +70,7 @@ export const NewWorksForYouRail: React.FC<NewWorksForYouRailProps & RailScrollPr
               title={title}
               onPress={() => {
                 trackEvent(tracks.tappedHeader())
-                navigate("/price-database")
+                navigate("/new-for-you")
               }}
             />
           </Flex>
